requirements.txt
# Core dependencies
neo4j>=5.14.0
python-dotenv>=1.0.0

# LangGraph and LLM dependencies (Gemini 2.0 Flash)
langgraph>=0.2.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.15
google-generativeai>=0.3.0
google-genai>=1.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Utilities
pydantic>=2.0.0
colorama>=0.4.6
rich>=13.0.0

# Web API dependencies
flask>=3.0.0
flask-cors>=4.0.0

# Vector Store dependencies (Phase 3.2)
chromadb>=1.5.0
pypdf>=6.0.0
sentence-transformers>=5.0.0

# Production server
gunicorn>=21.2.0

# Database adapters
psycopg2-binary>=2.9.9
redis>=5.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoring and logging
sentry-sdk>=1.40.0
python-json-logger>=2.0.7

# Rate limiting
flask-limiter>=3.5.0
